Respuestas a la encuesta de Psicotwitter

Cada fila es una respuesta por parte de una persona a la encuesta realizada por Alicia Franco (@AliciaFrancoXVE) en 2021.

clima

Format

Un dataset con 417 filas y 37 variables

id

Identificador de la respuesta

lugar

¿Desde dónde tuiteas normalmente? 1) España (o Europa), 2) Latinoamérica y Centroamérica, 3) Otro

genero

¿Cuál es tu género? 1) Mujer, 2) Hombre, 3) Persona no binaria, 4) Otro

psico1

¿Cuál es tu relación con la Psicología? 1) Interés, 2) Grado, 3) Posgrado, 4) Profesión

psico2

¿En qué medida te identificas con una corriente concreta en Psicología? Escala Likert 0 (Nada) - 5 (Completamente)

psico3

¿Cómo te sientes al interactuar con personas con ideas distintas a las tuyas sobre Psicología? Escala Likert 0 (Muy a disgusto) - 5 (Muy a gusto)

tw1

¿Cuánto tiempo llevas en Twitter? 1) Menos de 6 meses, 2) Menos de un año, 3) Menos de dos años, 4) Más de dos años

tw2

¿Con cuánta frecuencia usas Twitter? Aunque solo sea entrar a leer, no es necesario interactuar. 1) Casi nunca, 2) Una vez a la semana, 3) Varias veces a la semana, 4) Una vez al día, 5) Varias veces al día

tw_pers

¿En qué medida tu cuenta es personal? Escala Likert 0 (Nada) - 5 (Completamente)

tw_divulg

¿En qué medida tu cuenta es divulgativa? Escala Likert 0 (Nada) - 5 (Completamente)

tw_anonim

¿En qué medida tu cuenta es anónima? Escala Likert 0 (Nada) - 5 (Completamente)

tw_tuits

¿Con qué frecuencia interactúas en Twitter con tuits individuales? Escala Likert 0 (Nunca) - 5 (A diario)

tw_hilos_div

¿Con qué frecuencia interactúas en Twitter con tuits individuales? Escala Likert 0 (Nunca) - 5 (A diario)

tw_hilos_ot

¿Con qué frecuencia interactúas en Twitter con otro tipo de hilos? Escala Likert 0 (Nunca) - 5 (A diario)

tw_rt

¿Con qué frecuencia interactúas en Twitter con retuits? Escala Likert 0 (Nunca) - 5 (A diario)

tw_like

¿Con qué frecuencia interactúas en Twitter con likes? Escala Likert 0 (Nunca) - 5 (A diario)

tw_resp

¿Con qué frecuencia interactúas en Twitter con respuestas a tuits de otras cuentas? Escala Likert 0 (Nunca) - 5 (A diario)

exp_tw1

¿En qué medida crees que tu vida diaria está influenciada por lo que sucede en Twitter? Escala Likert 0 (Nada influenciada) - 5 (Completamente influenciada)

exp_tw2

¿Cómo te sientes en general en Twitter? Escala Likert 0 (Muy a disgusto) - 5 (Muy a gusto)

exp_tw3

¿Crees que el ambiente en Twitter (en tu experiencia personal) es más negativo o más positivo? Escala Likert 0 (Completamente negativo) - 5 (Completamente positivo)

exp_tw4

¿Crees que el ambiente en Twitter (en tu experiencia personal) es más competitivo o más cooperativo?) Escala Likert 0 (Completamente competitivo) - 5 (Completamente cooperativo)

comp1

¿Alguna vez has querido compartir algo en Twitter pero no lo has hecho por alguna razón?

comp2_verg

¿Por qué fue? Vergüenza

comp2_quediran

¿Por qué fue? Qué dirán

comp2_calidad

¿Por qué fue? Contenido de calidad

comp2_error

¿Por qué fue? Miedo a equivocarse

comp2_conf

¿Por qué fue? Evitar confrontación

comp2_nunca

¿Por qué fue? Nunca

comp3_divulg

¿Qué tipo de contenido era? Divulgación

comp3_apren

¿Qué tipo de contenido era? Algo que había aprendido

comp3_deb_cien

¿Qué tipo de contenido era? Debate científico

comp3_deb_poli

¿Qué tipo de contenido era? Debate político/personal

comp3_pers

¿Qué tipo de contenido era? Experiencia personal

comp3_nunca

¿Qué tipo de contenido era? Nunca

psico_tw1

¿Te consideras parte de "Psicotwitter"? 1) Sí, 2) Puede, 3) No, pero me gustaría, 4) No, y no me gustaría

psico_tw2

¿Qué te parece el término "Psicotwitter"? Escala Likert 0 (No me gusta nada) - 5 (Me encanta)

psico_tw3

¿Crees que el término "Psicotwitter" (en tu experiencia personal) es más negativo o más positivo? Escala Likert 0 (Completamente negativo) - 5 (Completamente positivo

Source

http://www.diamondse.info/

Author

Alicia Franco-Martínez

Examples

data("clima")